# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "FairPlay API",
    "version": "1.0.0",
    "description": "FairPlay REST API documentation"
  },
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Session key",
        "description": "Paste the sessionKey returned by /auth/login or /auth/verify-email."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "enum": [
              "BadRequest",
              "Conflict",
              "Forbidden",
              "InternalServerError",
              "InvalidJson",
              "NotFound",
              "PayloadTooLarge",
              "ServiceUnavailable",
              "TooManyRequests",
              "Unauthorized",
              "ValidationError"
            ]
          },
          "message": { "type": "string" }
        },
        "required": ["error", "message"]
      },
...
      "ApiOrValidationError": {
        "anyOf": [
          { "$ref": "#/components/schemas/ApiError" },
          { "$ref": "#/components/schemas/ValidationError" }
        ]
      },
...
      "PublicVideosResponse": {
        "type": "object",
        "properties": {
          "videos": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "publicId": {
                  "type": "string",
                  "pattern": "^[A-Za-z0-9_]{10}$",
                  "example": "AbCdEf123_"
                },
                "title": { "type": "string", "example": "Me at the zoo" },
                "createdAt": {
                  "type": "string",
                  "format": "date-time",
                  "example": "2005-04-23T20:31:52.000Z"
                },
                "thumbnailPath": {
                  "type": "string",
                  "nullable": true,
                  "pattern": "^\\/(?!\\/)[^\\s]*$",
                  "example": "/videos/AbCdEf123_/thumbnail"
                },
                "creator": {
                  "type": "object",
                  "properties": {
                    "username": { "type": "string", "example": "fairplay_creator" },
                    "displayName": {
                      "type": "string",
                      "nullable": true,
                      "example": "FairPlay Creator"
                    }
                  },
                  "required": ["username", "displayName"]
                },
                "viewCount": { "type": "integer", "minimum": 0, "example": 128 },
                "duration": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": true,
                  "example": 19
                }
              },
              "required": [
                "publicId",
                "title",
                "createdAt",
                "thumbnailPath",
                "creator",
                "viewCount",
                "duration"
              ]
            }
          },
          "total": { "type": "integer", "minimum": 0, "example": 42 },
          "nextCursor": {
            "type": "object",
            "nullable": true,
            "properties": {
              "createdAt": {
                "type": "string",
                "format": "date-time",
                "example": "2026-01-01T00:00:00.000Z"
              },
              "publicId": {
                "type": "string",
                "pattern": "^[A-Za-z0-9_]{10}$",
                "example": "AbCdEf123_"
              }
            },
            "required": ["createdAt", "publicId"]
          }
        },
        "required": ["videos", "total", "nextCursor"]
      },
...
      "VideoRatingAggregateResponse": {
        "type": "object",
        "properties": {
          "ratingAverage": {
            "type": "number",
            "minimum": 0,
            "maximum": 5,
            "description": "Arithmetic mean rounded to one decimal place, or zero when unrated",
            "example": 4.3
          },
          "ratingCount": { "type": "integer", "minimum": 0, "example": 12 }
        },
        "required": ["ratingAverage", "ratingCount"]
      },
...
    },
...
  },
  "paths": {
...
    "/videos": {
      "get": {
        "operationId": "listPublicVideos",
        "summary": "List the public video feed",
        "tags": ["Videos"],
        "security": [],
        "parameters": [
          {
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "example": 20 },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-01-01T00:00:00.000Z"
            },
            "required": false,
            "name": "cursorCreatedAt",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_]{10}$",
              "example": "AbCdEf123_"
            },
            "required": false,
            "name": "cursorPublicId",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated public video feed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PublicVideosResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiOrValidationError" }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          }
        }
      },
      "post": {
        "operationId": "createVideo",
        "summary": "Create video metadata",
        "description": "Creates draft video metadata. The optional allowComments preference defaults to true and cannot be changed through this API version after creation.",
        "tags": ["Videos"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/CreateVideoRequest" } }
          }
        },
        "responses": {
          "201": {
            "description": "Video metadata created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CreateVideoResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiOrValidationError" }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          }
        }
      }
    },
    "/videos/search": {
      "get": {
        "operationId": "searchPublicVideos",
        "summary": "Search public videos and creators",
        "tags": ["Videos"],
        "security": [],
        "parameters": [
          {
            "schema": { "type": "integer", "minimum": 1, "maximum": 100, "example": 20 },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "date-time",
              "example": "2026-01-01T00:00:00.000Z"
            },
            "required": false,
            "name": "cursorCreatedAt",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_]{10}$",
              "example": "AbCdEf123_"
            },
            "required": false,
            "name": "cursorPublicId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "minLength": 2,
              "maxLength": 254,
              "example": "launch recap",
              "description": "Runs two independent searches: videos match case-insensitive literal title/description substrings or an exact tag, while creators match case-insensitive literal username/display-name substrings. Creator matches never add that creator's videos to the videos section."
            },
            "required": true,
            "description": "Runs two independent searches: videos match case-insensitive literal title/description substrings or an exact tag, while creators match case-insensitive literal username/display-name substrings. Creator matches never add that creator's videos to the videos section.",
            "name": "search",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "enum": ["newest", "oldest"],
              "example": "newest",
              "description": "Sort by creation time. Defaults to newest."
            },
            "required": false,
            "description": "Sort by creation time. Defaults to newest.",
            "name": "sort",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated public video results with bounded creator matches",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/PublicVideoSearchResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiOrValidationError" }
              }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          }
        }
      }
    },
...
    "/videos/{publicId}/rating": {
      "get": {
        "operationId": "getVideoRating",
        "summary": "Get the public rating aggregate for a video",
        "tags": ["Videos"],
        "security": [],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_]{10}$",
              "example": "AbCdEf123_"
            },
            "required": true,
            "name": "publicId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Video rating aggregate",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VideoRatingAggregateResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiOrValidationError" }
              }
            }
          },
          "404": {
            "description": "Video not found or inaccessible",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          }
        }
      },
      "put": {
        "operationId": "rateVideo",
        "summary": "Create or update the current user rating for a video",
        "tags": ["Videos"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_]{10}$",
              "example": "AbCdEf123_"
            },
            "required": true,
            "name": "publicId",
            "in": "path"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/RateVideoRequest" } }
          }
        },
        "responses": {
          "200": {
            "description": "Updated video rating aggregate",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/VideoRatingResponse" }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiOrValidationError" }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "403": {
            "description": "Video owners cannot rate their own videos",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "404": {
            "description": "Video not found or inaccessible",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "429": {
            "description": "Too many requests",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          },
          "503": {
            "description": "Video rating temporarily unavailable",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
            }
          }
        }
      }
    },
...
  }
}